Update to Opencore 1.0.0
Unfortunately, in the last few months, I haven't had a lot of time to dedicate to the development of the project. I updated Opencore and kexts to the latest version available. There are also some minor differences in the ACPI patches. I didn't have a lot of time to test, but everything seems fine. This also closes issue 9 (https://github.com/antoniomcr96/Thinkpad-L390-Yoga-macOS-Opencore/issues/9) for the moment.